Add support for cries/sprites
This would previously point to raw.githubusercontent.com, which is
problematic as:

 - scraping this might result in getting rate-limited by GitHub.
 - the path may not exist in the future.

We thus incorporate this into the image (increasing its size by over
1GB) to insulate ourselves from this.

Some quick experimentation with compression indicates that we shouldn't
bother trying to compress these png/ogg/gif files - these are already
optimised so this does nothing. However, it might be worth experimenting
with compressing the json files.