Migrate to scrapy using uv
Our shell scripts are getting a bit out of hand; instead, we migrate to
scrapy which is far better-suited to scraping an endpoint with better
logging and tuning available.

This shouldn't really change how long the build takes; the bottleneck
appears to be `pokeapi`, not the actual operations performed after we
have downloaded a response.